Collision-fix (#2)
Due to me mainly having ignored this script for the past weeks and mostly only working with one certain and big test model, I didn't notice that the script fucks your collision objects up. That is, it doesn't matter what collision-properties you give them, the script will just reset them to the default values while it is running. This made using it on small models practically impossible. Fixed that with 1.30.1